MSFT_PackageManagement: Fix conversion of AdditionalParameters for nondefault providers
Providers which are not included in PackageManagement, such as
DockerMsftProvider, are not loaded automatically. As a consequence,
their dynamic parameters are not returned by Get-Command. The fix
is to make sure the provider is loaded, which can be done if the
ProviderName is specified (always a good idea).

Also, in order to improve the fidelity of parameter type detection,
it is better to look only at parameters exposed by that provider.
Conveniently, for each provider, a separate ParameterSet is defined,
named after that provider.